Updated the code
The code now works for from one to three digit number. I can now make it work for the other numbers. I also need to make it more of a memory test by making the output disappear. This seems difficult to do on Colab from the previous tests we have done this project but I will continue looking into it. If not I will need to get rid of the output manually by outputting spaces.